refactor slippage
