+++ components/Dao/Bancor/Earn.js
@@ -11,7 +11,7 @@ import { useLayoutStore } from "../../../stores/useLayoutStore"
 import { usePlaygroundStore } from "../../../stores/usePlaygroundStore"
 
 import TokenInput from "../TokenInput"
-import Slippage   from '../Uniswap/Slippage'
+import Slippage   from '../Slippage'
 
 const Earn = ({ ethToken }) => {
   const signer = useLayoutStore(state => state.signer)
